+++ .github/workflows/node.js.yaml
@@ -1,0 +1,25 @@
+name: Node.js CI
+
+on:
+  push:
+    branches: main
+  pull_request:
+    branches: main
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+      - uses: pnpm/action-setup@v2
+        with:
+          version: latest
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 21
+          cache: pnpm
+      - run: pnpm install --frozen-lockfile
+      - run: pnpm run build --noEmit
+      - run: pnpm run lint
+      - run: pnpm run test
