+++ .github/workflows/pnpm-version-patch.yaml
@@ -1,0 +1,44 @@
+name: Node.js Patch
+
+on: workflow_dispatch
+
+jobs:
+  patch:
+    runs-on: ubuntu-latest
+
+    permissions:
+      contents: write
+
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          ssh-key: "${{ secrets.DEPLOY_KEY_PRIVATE }}"
+      - uses: pnpm/action-setup@v2
+        with:
+          version: latest
+      - uses: actions/setup-node@v3
+        with:
+          cache: pnpm
+          node-version: 21
+      - name: Configure Git, patch, release and push
+        run: |
+          git config user.name 'github-actions[bot]'
+          git config user.email '41898282+github-actions[bot]@users.noreply.github.com'
+
+          git config commit.gpgsign true
+          git config gpg.format ssh
+          git config user.signingkey 'key::${{ vars.DEPLOY_KEY_PUBLIC }}'
+
+          eval `ssh-agent -s`
+          ssh-add - <<< '${{ secrets.DEPLOY_KEY_PRIVATE }}'
+
+          VERSION=$(pnpm version patch --no-git-tag-version)
+
+          git commit --all --message "🔖 $VERSION"
+          git tag --annotate --message "🔖 $VERSION" --sign $VERSION
+          git push
+          git push --tags
+
+          gh release create $VERSION --generate-notes --title $VERSION --verify-tag
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
