+++ .npmrc
@@ -1,0 +1,1 @@
+resolution-mode=highest
