+++ .prettierignore
@@ -1,0 +1,5 @@
+dist/
+docs/
+node_modules/
+package-lock.json
+pnpm-lock.yaml
