+++ .vscode/extensions.json
@@ -1,0 +1,7 @@
+{
+	"recommendations": [
+		"dbaeumer.vscode-eslint",
+		"esbenp.prettier-vscode",
+		"vitest.explorer"
+	]
+}
