+++ package.json
@@ -1,46 +1,49 @@
 {
-  "name": "openai-chat-tokens",
-  "version": "0.2.8",
-  "description": "Estimate the number of tokens an OpenAI chat completion request will use",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "test": "vitest",
-    "build": "tsc",
-    "prepublishOnly": "npm run build",
-    "lint": "prettier --check src tests"
-  },
-  "author": {
-    "email": "harry@hmarr.com",
-    "name": "Harry Marr",
-    "url": "https://hmarr.com"
-  },
-  "repository": {
-    "url": "https://github.com/hmarr/openai-chat-tokens",
-    "type": "git"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "openai": "^4.18.0",
-    "prettier": "^3.1.0",
-    "typescript": "^5.2.2",
-    "vite": "^5.1.6",
-    "vitest": "^1.4.0"
-  },
-  "dependencies": {
-    "js-tiktoken": "^1.0.7"
-  },
-  "prettier": {
-    "trailingComma": "all"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    }
-  },
-  "type": "module"
+	"name": "@coderabbitai/openai-chat-tokens",
+	"version": "0.2.8",
+	"description": "Estimate the number of tokens an OpenAI chat completion request will use",
+	"homepage": "https://github.com/coderabbitai/openai-chat-tokens",
+	"bugs": {
+		"url": "https://github.com/coderabbitai/openai-chat-tokens/issues"
+	},
+	"license": "MIT",
+	"files": [
+		"dist",
+		"!dist/**/*.test.*"
+	],
+	"main": "dist/index.js",
+	"module": "dist/index.js",
+	"repository": "github:coderabbitai/openai-chat-tokens",
+	"scripts": {
+		"build": "tsc",
+		"docs": "typedoc",
+		"format": "prettier --write .",
+		"lint": "eslint . && prettier --check .",
+		"lint:fix": "eslint --fix .; prettier --list-different --write .",
+		"test": "vitest",
+		"test:watch": "vitest --watch"
+	},
+	"devDependencies": {
+		"@typescript-eslint/eslint-plugin": "^7.4.0",
+		"@typescript-eslint/parser": "^7.4.0",
+		"eslint": "^8.57.0",
+		"eslint-config-prettier": "^9.1.0",
+		"openai": "^4.18.0",
+		"prettier": "^3.1.0",
+		"typedoc": "^0.25.12",
+		"typescript": "^5.2.2",
+		"vite": "^5.1.6",
+		"vitest": "^1.4.0"
+	},
+	"dependencies": {
+		"js-tiktoken": "^1.0.7"
+	},
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"default": "./dist/index.js"
+		}
+	},
+	"types": "dist/index.d.ts",
+	"type": "module"
 }
