+++ typedoc.json
@@ -1,0 +1,4 @@
+{
+	"$schema": "https://typedoc.org/schema.json",
+	"entryPoints": ["./src/index.ts"]
+}
